Sprinkler/spray irrigation is the method
Sprinkler/spray irrigation is the method of applying water to a controlled manner in that is similar to rainfall. The water is distributed through a network that may consist of pumps, valves, pipes, and sprinklers. Irrigation sprinklers can be used for residential, industrial, and agricultural usage.Spray irrigation is a modern and commonly-used system of irrigating, but it also requires machinery. This system is similar to the way you might water your lawn at home - stand there with a hose and spray the water out in all directions. Large scale spray irrigation systems are in use on large farms today. The systems can simply be long hoses with sprinklers along the length or a center-pivot system that traverses a circle in the fields. This picture shows a simple sprinkler irrigation system—it is one of the less-efficient models because a large part of the water shot into the air evaporates. 

Arduino is used to control the motor thrust via motor driver and PWM