Fan On Delay Timer
Intelligent Induction: the load will be opened automatically as person comes; the load will be off automatically as a person leaves. Intelligent Delay: the switch will automatically add a delay time period after detecting every move of a person. It will regard the last move as a new start point for delay time